+++ semTools/R/efa.R
@@ -1,5 +1,6 @@
 ### Sunthud Pornprasertmanit & Terrence D. Jorgensen
 ### Last updated: 27 May 2020
+###   DEPRECATED  24 November 2022
 ### fit and rotate EFA models in lavaan
 
 
@@ -61,9 +62,35 @@
 ##'                   z = rnorm(nrow(HolzingerSwineford1939), 0, 1))
 ##' unrotated2 <- efaUnrotate(dat, nf = 2, varList = paste0("x", 1:9), aux = "z")
 ##'
+##' @name efaUnrotate-deprecated
+##' @usage
+##' efaUnrotate(data = NULL, nf, varList = NULL,
+##'             start = TRUE, aux = NULL, ...)
+##' @seealso \code{\link{semTools-deprecated}}
+##' @keywords internal
+NULL
+
+
+##' @rdname semTools-deprecated
+##'
+##' @section Support functions for exploratory factor analysis in \code{lavaan}:
+##' The \code{efaUnrotate}, \code{orthRotate}, \code{oblqRotate}, and
+##' \code{funRotate} functions will no longer be supported. These functions
+##' allowed users to estimate EFA parameters with \code{lavaan}. Instead, users
+##' can now directly use \code{lavaan}'s \code{\link[lavaan]{efa}} function.
+##' Exploratory SEM (ESEM) is also supported by lavaan using special operators
+##' in \code{lavaan}'s \code{\link[lavaan]{model.syntax}}; see
+##' \url{https://github.com/yrosseel/lavaan/issues/112} for details.
+##'
 ##' @export
 efaUnrotate <- function(data = NULL, nf, varList = NULL,
                         start = TRUE, aux = NULL, ...) {
+
+  .Deprecated(msg = c("The efaUnrotate function is deprecated, and it will ",
+                      "cease to be included in future versions of semTools. ",
+                      "The lavaan package itself now provides support EFA. ",
+                      "See help('semTools-deprecated) for details."))
+
   efaArgs <- list(...)
   if (is.null(data)) {
     ## check for summary statistics
@@ -332,8 +359,29 @@ setMethod("summary", signature(object = "EFA"),
 ##' funRotate(unrotated, fun = "targetQ", Target = target)
 ##' }
 ##'
+##' @name rotate-deprecated
+##'
+##' @usage
+##' orthRotate(object, method = "varimax", ...)
+##'
+##' oblqRotate(object, method = "quartimin", ...)
+##'
+##' funRotate(object, fun, ...)
+##'
+##' @seealso \code{\link{semTools-deprecated}}
+##' @keywords internal
+NULL
+
+
+##' @rdname semTools-deprecated
 ##' @export
 orthRotate <- function(object, method = "varimax", ...) {
+
+  .Deprecated(msg = c("The orthRotate function is deprecated, and it will ",
+                      "cease to be included in future versions of semTools. ",
+                      "The lavaan package itself now provides support EFA. ",
+                      "See help('semTools-deprecated) for details."))
+
 	requireNamespace("GPArotation")
 	if (!("package:GPArotation" %in% search())) attachNamespace("GPArotation")
 	mc <- match.call()
@@ -357,10 +405,16 @@ orthRotate <- function(object, method = "varimax", ...) {
 
 
 
-##' @rdname rotate
+##' @rdname semTools-deprecated
 ##' @export
 oblqRotate <- function(object, method = "quartimin", ...) {
-	requireNamespace("GPArotation")
+
+  .Deprecated(msg = c("The oblqRotate function is deprecated, and it will ",
+                      "cease to be included in future versions of semTools. ",
+                      "The lavaan package itself now provides support EFA. ",
+                      "See help('semTools-deprecated) for details."))
+
+  requireNamespace("GPArotation")
 	if (!("package:GPArotation" %in% search())) attachNamespace("GPArotation")
 	mc <- match.call()
 	initL <- getLoad(object)
@@ -383,10 +437,16 @@ oblqRotate <- function(object, method = "quartimin", ...) {
 
 
 
-##' @rdname rotate
+##' @rdname semTools-deprecated
 ##' @export
 funRotate <- function(object, fun, ...) {
-	stopifnot(is.character(fun))
+
+  .Deprecated(msg = c("The funRotate function is deprecated, and it will ",
+                      "cease to be included in future versions of semTools. ",
+                      "The lavaan package itself now provides support EFA. ",
+                      "See help('semTools-deprecated) for details."))
+
+  stopifnot(is.character(fun))
 	requireNamespace("GPArotation")
 	if (!("package:GPArotation" %in% search())) attachNamespace("GPArotation")
 	mc <- match.call()